fix: split multicron into both include and plugin components
