+++ orbiter/assets/timetables/multi_cron/multi_cron_timetable_plugin_src.py
@@ -1,0 +1,7 @@
+from airflow.plugins_manager import AirflowPlugin
+from includes.multi_cron_timetable import MultiCronTimetable
+
+
+class MultiCronTimetablePlugin(AirflowPlugin):
+    name = "multi_cron_timetable"
+    timetables = [MultiCronTimetable]
