+++ tests/orbiter/objects/project_test.py
@@ -9,6 +9,7 @@ from orbiter.objects.env_var import OrbiterEnvVar
 from orbiter.objects.pool import OrbiterPool
 from orbiter.objects.project import OrbiterProject
 from orbiter.objects.task import OrbiterTask
+from orbiter.objects.timetables.multi_cron_timetable import OrbiterMultiCronTimetable
 from orbiter.objects.variable import OrbiterVariable
 
 
@@ -17,7 +18,12 @@ def test_project_render(tmpdir):
     # noinspection PyArgumentList
     project = OrbiterProject().add_dags(
         dags=[
-            OrbiterDAG(dag_id="foo", file_path="foo.py", schedule=None, doc_md="foo").add_tasks(
+            OrbiterDAG(
+                dag_id="foo",
+                file_path="foo.py",
+                schedule=OrbiterMultiCronTimetable(cron_defs=["0 1 * * *", "*/5 0 * * *"]),
+                doc_md="foo",
+            ).add_tasks(
                 tasks=[
                     OrbiterTask(
                         task_id="foo",
@@ -46,7 +52,7 @@ def test_project_render(tmpdir):
     project.render(tmpdir)
 
     actual_requirements = (tmpdir / "requirements.txt").read_text()
-    expected_requirements = "apache-airflow\npendulum"
+    expected_requirements = "apache-airflow\ncroniter\npendulum"
     assert actual_requirements == expected_requirements
 
     actual_packages = (tmpdir / "packages.txt").read_text()
@@ -77,7 +83,18 @@ def test_project_render(tmpdir):
     actual_dag = actual_dag.read_text()
     expected_dag = """from airflow import DAG
 from airflow.operators.empty import EmptyOperator
+from include.multi_cron_timetable import MultiCronTimetable
 from pendulum import DateTime, Timezone
-with DAG(dag_id='foo', schedule=None, start_date=DateTime(1970, 1, 1, 0, 0, 0), catchup=False, doc_md='foo'):
+with DAG(dag_id='foo', schedule=MultiCronTimetable(cron_defs=['0 1 * * *', '*/5 0 * * *']), start_date=DateTime(1970, 1, 1, 0, 0, 0), catchup=False, doc_md='foo'):
     foo_task = EmptyOperator(task_id='foo', doc='some other thing')"""
     assert actual_dag == expected_dag
+
+    actual_include = tmpdir / "include/multi_cron_timetable.py"
+    assert actual_include.exists(), actual_include
+    actual_include = actual_include.read_text()
+    assert "class MultiCronTimetable(Timetable):" in actual_include
+
+    actual_plugin = tmpdir / "plugins/multi_cron_timetable.py"
+    assert actual_plugin.exists(), actual_plugin
+    actual_plugin = actual_plugin.read_text()
+    assert "class MultiCronTimetablePlugin(AirflowPlugin)" in actual_plugin
